requirements.txt
requests
python-dotenv
openai
google-auth-oauthlib
google-auth-httplib2
